added working example for web
