+++ example/lib/unknown_page.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class UnknownPage extends StatelessWidget {
+  const UnknownPage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+        child: Text("Wrong page"),
+      ),
+    );
+  }
+}
